update for linux.

+++ setup.sh
@@ -72,8 +72,9 @@ exec_install() {
 
   if [ ! -e $HOME/.rc-org ]; then
     mkdir $HOME/.rc-org
-    for F in "$BACKUP_FILES" ;do
-      if [ -e "${HOME}/${F}" ]; then
+    for F in $BACKUP_FILES ;do
+      echo ${HOME}/$F
+      if [ -e "${HOME}/${F}" -a ! -L "${HOME}/${F}" ]; then
         echo -e "${RED}mv $HOME/$F $HOME/.rc-org${DEFAULT}"
         mv $HOME/$F $HOME/.rc-org
       fi
